issue: move zip functions to package issue
Since we are not using zip functions anywhere else, let's move zip
functions to where they are used, and also unexport all of them.
